EasyTasks/src/FindAllNumbersDisappearedInAnArray/Solution.java
package FindAllNumbersDisappearedInAnArray;

import java.util.ArrayList;
import java.util.List;

public class Solution {
    public List<Integer> findDisappearedNumbers(int[] nums) {
        int count = 0;
        for (int i = 0; i < nums.length; i++) {
            int ind = Math.abs(nums[i]) - 1;
            if (nums[ind] > 0) {
                nums[ind] = -nums[ind];
                count++;
            }
        }
        List<Integer> result = new ArrayList<>(nums.length - count);
        for (int i = 0; i < nums.length; i++) {
            if (nums[i] > 0) {
                result.add(i + 1);
            }
        }
        return result;
    }
}
